Documentation and minor refactoring of middleware code
