+++ src/middleware/marshalling/dccl.h
@@ -23,7 +23,7 @@
 #ifndef SerializeParseDCCL20190717H
 #define SerializeParseDCCL20190717H
 
-#include <dccl/codec.h>
+#include "detail/dccl_serializer_parser.h"
 
 #include "protobuf.h"
 
@@ -34,112 +34,15 @@ namespace middleware
 namespace protobuf
 {
 class InterVehicleSubscription;
-class DCCLForwardedData;
 } // namespace protobuf
 
-/// \brief Wraps a dccl::Codec in a thread-safe way to make it usable by SerializerParserHelper
-struct DCCLSerializerParserHelperBase
-{
-  private:
-    static std::unique_ptr<dccl::Codec> codec_;
-
-  protected:
-    static std::mutex dccl_mutex_;
-
-    struct LoaderBase
-    {
-        LoaderBase() = default;
-        virtual ~LoaderBase() = default;
-    };
-
-    template <typename DataType> struct Loader : public LoaderBase
-    {
-        Loader() { codec().load<DataType>(); }
-        ~Loader() { codec().unload<DataType>(); }
-    };
-
-    struct LoaderDynamic : public LoaderBase
-    {
-        LoaderDynamic(const google::protobuf::Descriptor* desc) : desc_(desc)
-        {
-            codec().load(desc_);
-        }
-        ~LoaderDynamic() { codec().unload(desc_); }
-
-      private:
-        const google::protobuf::Descriptor* desc_;
-    };
-
-    static std::unordered_map<const google::protobuf::Descriptor*, std::unique_ptr<LoaderBase>>
-        loader_map_;
-
-    template <typename DataType> static void check_load()
-    {
-        const auto* desc = DataType::descriptor();
-        if (!loader_map_.count(desc))
-            loader_map_.insert(
-                std::make_pair(desc, std::unique_ptr<LoaderBase>(new Loader<DataType>())));
-    }
-
-    static void check_load(const google::protobuf::Descriptor* desc)
-    {
-        if (!loader_map_.count(desc))
-            loader_map_.insert(
-                std::make_pair(desc, std::unique_ptr<LoaderBase>(new LoaderDynamic(desc))));
-    }
-
-    static dccl::Codec& codec()
-    {
-        if (!codec_)
-            codec_.reset(new dccl::Codec);
-        return *codec_;
-    }
-
-    static dccl::Codec& set_codec(dccl::Codec* new_codec)
-    {
-        codec_.reset(new_codec);
-        loader_map_.clear();
-        return *new_codec;
-    }
-
-  public:
-    DCCLSerializerParserHelperBase() = default;
-    virtual ~DCCLSerializerParserHelperBase() = default;
-
-    template <typename CharIterator> static unsigned id(CharIterator begin, CharIterator end)
-    {
-        std::lock_guard<std::mutex> lock(dccl_mutex_);
-        return codec().id(begin, end);
-    }
-
-    static unsigned id(const std::string full_name)
-    {
-        std::lock_guard<std::mutex> lock(dccl_mutex_);
-        auto* desc = dccl::DynamicProtobufManager::find_descriptor(full_name);
-        if (desc)
-            return codec().id(desc);
-        else
-            return 0;
-    }
-
-    static void
-    load_forwarded_subscription(const goby::middleware::intervehicle::protobuf::Subscription& sub);
-    static goby::middleware::intervehicle::protobuf::DCCLForwardedData
-    unpack(const std::string& bytes);
-
-    static void load_library(const std::string& library)
-    {
-        std::lock_guard<std::mutex> lock(dccl_mutex_);
-        codec().load_library(library);
-    }
-};
 
 /// \brief Specialization for DCCL message types that are fully qualified Protobuf message types (static), e.g. DataType == Foo for "message Foo"
 ///
 /// DCCL is defined as distinct from the PROTOBUF Marshalling scheme if the DataTYpe has the DCCLParameters enumeration, as output by the \c protoc-gen-dccl plugin to \c protoc. If this plugin isn't used when compiling your .proto files, DCCL types will be identified as Protobuf types.
 template <typename DataType>
 struct SerializerParserHelper<DataType, MarshallingScheme::DCCL>
-    : public DCCLSerializerParserHelperBase
+    : public detail::DCCLSerializerParserHelperBase
 {
   public:
     /// \brief Serialize message using DCCL encoding
@@ -200,7 +103,7 @@ struct SerializerParserHelper<DataType, MarshallingScheme::DCCL>
 /// \brief Specialization for runtime introspection of DCCL messages using google::protobuf::Message base class (works for publish and subscribe_regex only)
 template <>
 struct SerializerParserHelper<google::protobuf::Message, MarshallingScheme::DCCL>
-    : public DCCLSerializerParserHelperBase
+    : public detail::DCCLSerializerParserHelperBase
 {
   public:
     /// Serialize DCCL/Protobuf message (using DCCL encoding)
