+++ src/middleware/src.cmake
@@ -16,7 +16,7 @@ set(MIDDLEWARE_SRC
   middleware/marshalling/interface.cpp
   middleware/marshalling/dccl.cpp 
   middleware/transport/interthread.cpp
-  middleware/intervehicle/driver-thread.cpp
+  middleware/transport/intervehicle/driver_thread.cpp
   middleware/application/configuration_reader.cpp
   middleware/log/log_entry.cpp
   ${MIDDLEWARE_PROTO_SRCS} ${MIDDLEWARE_PROTO_HDRS}
